+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added a custom Checkbox component
+* Set 'caret' as a default iconClass for "switchers"
 
 ## 1.1.1
 * Changed disabling logic a bit (removed disableCheckboxes prop and added disabled prop)